Update styles/pbtech/theme/pbwow.css
To make viewing of member lists more readable, the hover:bg color of the rows
was changed to the same hover:bg color of the forum topics and sections (white).
Previously black, the color makes it hard to view rows of users in group
"Registered Users" as their group color is also black.